requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
numpy==1.24.3
faiss-cpu==1.7.4
python-dotenv==1.0.0
gunicorn==20.1.0

# RAG + LLM Dependencies
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.35.2
accelerate==0.25.0
